Improve documentation about local development
To hack on a source repository, one should always do it inside cerbero's shell, since it ensures that you will be using the correct environment (in case, for example, autotools triggers an autoreconf) and it also exposes the tools already installed in the prefix. In the example so you can use gst-inspect-1.0.exe directly, which is more convinient if for example you are in the gst-plugins-bad sources instead of gstreamer.